backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
gunicorn==21.2.0

# AI (API-based, lightweight)
google-generativeai==0.8.5

# Data & similarity
numpy==2.3.1
scikit-learn==1.5.0

# File handling
python-docx==1.2.0
Pillow==11.1.0
PyMuPDF==1.26.1
pdf2image==1.16.3